requirements.txt
pandas==1.2.3
tabula-py==2.2.0
